Add files via upload
Each HTML is now sending data to each one. One last thing we need to do is distribute the JSON evenly to each HTML so each HTML doesn't load the same JSON file which causes data to be repeating. Another thing we can do is have it working when all four clients(HTML) is loaded and then start reading to each screen. But we don't have to.